Module rmc_beaglebone_core_2018_2019::robot_map

The robot map is a file filled with key constants such as pin numbers and network ports that may change over time. It is used to make reconfiguring pinouts a simpler process.

Enums

MotorID

Constants

ADDRESS
The address for the tcp server used to communicate with the driver station. Zero indicates that the server will accept connections from any IP.

FRONT_LEFT_DIRECTION

FRONT_LEFT_PWM_CHIP

FRONT_LEFT_PWM_NUMBER

FRONT_RIGHT_DIRECTION

FRONT_RIGHT_PWM_CHIP

FRONT_RIGHT_PWM_NUMBER

LOG_PATH
The path of the folder which logs will be kept in. If the folder does not exist, the program will create it.

PORT
The port used for communicating with the driver station.

REAR_LEFT_DIRECTION

REAR_LEFT_PWM_CHIP

REAR_LEFT_PWM_NUMBER

REAR_RIGHT_DIRECTION

REAR_RIGHT_PWM_CHIP

REAR_RIGHT_PWM_NUMBER